-fix vite production builds -remove deps unused or unwanted -remove date service for now -remove timespan for now -fix runtime up
