+++ index.html
@@ -7,15 +7,6 @@
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lato" />
   <base href="/" />
-  <script>window.global = window;</script>
-  <script type="module">
-    import process from "process";
-    import { Buffer } from "buffer";
-    import EventEmitter from "events";
-    window.Buffer = Buffer;
-    window.process = process;
-    window.EventEmitter = EventEmitter;
-  </script>
 </head>
 
 <body>
