+++ package.json
@@ -27,11 +27,11 @@
     "start-celo": "cross-env KOL_NETWORK=Celo vite",
     "start-alfajores": "vite",
     "build": "npm run build-celo",
-    "build-celo": "cross-env DOTENV_CONFIG_PATH=.env.production KOL_NETWORK=Celo rimraf dist && webpack --env production",
-    "build-celo-dev": "cross-env DOTENV_CONFIG_PATH=.env.production KOL_NETWORK=Celo rimraf dist && webpack --env development",
-    "build-alfajores": "cross-env DOTENV_CONFIG_PATH=.env.production KOL_NETWORK=Alfajores rimraf dist && webpack --env production",
+    "build-celo": "cross-env DOTENV_CONFIG_PATH=.env.production KOL_NETWORK=Celo rimraf dist && vite build",
+    "build-celo-dev": "cross-env DOTENV_CONFIG_PATH=.env.production KOL_NETWORK=Celo rimraf dist && vite build --env development",
+    "build-alfajores": "cross-env DOTENV_CONFIG_PATH=.env.production KOL_NETWORK=Alfajores rimraf dist && vite build",
     "build-alfajores-dev": "cross-env DOTENV_CONFIG_PATH=.env.production KOL_NETWORK=Alfajores rimraf dist && webpack --env development",
-    "analyze": "cross-env DOTENV_CONFIG_PATH=.env.production rimraf dist && webpack --env production --analyze",
+    "analyze": "cross-env DOTENV_CONFIG_PATH=.env.production rimraf dist && vite build --analyze",
     "fetchContracts": "node scripts/fetchContracts.js",
     "test": "vitest -r src/",
     "unit": "jest",
@@ -44,16 +44,15 @@
     "vercel:alias": "vercel alias set $npm_config_deployment $npm_config_branch --token $npm_config_token --scope $npm_config_scope"
   },
   "dependencies": {
+    "axios": "^0.27.2",
+    "rxjs": "^7.5.6",
     "@aurelia/i18n": "dev",
     "@aurelia/validation-html": "dev",
     "@celo-tools/celo-ethers-wrapper": "^0.3.0",
     "@material-design-icons/svg": "^0.11.4",
     "@metamask/detect-provider": "^1.2.0",
-    "@microsoft/fast-components": "^2.30.6",
-    "@microsoft/fast-element": "^1.10.2",
     "@walletconnect/web3-provider": "^1.7.8",
     "aurelia": "dev",
-    "aurelia-fast-adapter": "^1.0.5",
     "browserify-zlib": "^0.2.0",
     "buffer": "^6.0.3",
     "chart.js": "^3.8.0",
@@ -61,33 +60,30 @@
     "ethers": "^5.6.9",
     "events": "^3.3.0",
     "i18next-intervalplural-postprocessor": "^3.0.0",
-    "moment-timezone": "^0.5.34",
     "process": "^0.11.10",
     "stream-browserify": "^3.0.0",
     "util": "^0.10.4",
     "web3modal": "^1.9.8"
   },
   "devDependencies": {
+    "@esbuild-plugins/node-globals-polyfill": "^0.1.1",
+    "rollup-plugin-polyfill-node": "^0.10.2",
+    "rollup-plugin-visualizer": "^5.8.0",
     "@rollup/plugin-inject": "^4.0.4",
+    "rollup-pluginutils": "^2.8.2",
+    "vite-tsconfig-paths": "^3.5.0",
     "@aurelia/testing": "dev",
     "@aurelia/ts-jest": "dev",
-    "@aurelia/webpack-loader": "dev",
     "@playwright/test": "^1.22.2",
     "@types/jest": "^28.1.1",
     "@types/node": "^17.0.35",
     "@typescript-eslint/eslint-plugin": "^5.27.1",
     "@typescript-eslint/parser": "^5.27.1",
-    "autoprefixer": "^10.4.7",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.7.1",
-    "cypress": "^9.6.1",
-    "cypress-shadow-dom": "^1.4.1",
-    "dotenv-webpack": "^7.1.0",
     "ethereum-abi-types-generator": "^1.3.2",
     "eslint": "^8.17.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-typescript": "^2.7.1",
-    "eslint-import-resolver-webpack": "^0.13.2",
     "eslint-plugin-cypress": "^2.12.1",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-prettier": "^4.2.1",
@@ -97,7 +93,6 @@
     "glob": "^8.0.3",
     "happy-dom": "^6.0.4",
     "html-loader": "^4.1.0",
-    "html-webpack-plugin": "^5.5.0",
     "htmlhint": "^1.1.4",
     "http-browserify": "1.7.0",
     "https-browserify": "^1.0.0",
@@ -106,50 +101,17 @@
     "jest-transform-stub": "^2.0.0",
     "postcss": "^8.4.14",
     "postcss-loader": "^7.0.0",
-    "rimraf": "^3.0.2",
     "rollup-plugin-html": "^0.2.1",
-    "sass": "^1.52.0",
-    "sass-loader": "^13.0.0",
-    "start-server-and-test": "^1.14.0",
-    "style-loader": "^3.3.1",
     "stylelint": "^14.9.1",
     "stylelint-config-prettier": "^9.0.3",
     "stylelint-config-standard": "^26.0.0",
     "stylelint-config-standard-scss": "^5.0.0",
     "stylelint-prettier": "^2.0.0",
-    "ts-loader": "^9.3.0",
     "tslib": "^2.4.0",
     "typescript": "^4.6.4",
     "vercel": "^25.1.0",
     "vite": "latest",
     "vitest": "^0.21.1",
-    "vitest-mock-extended": "^0.1.13",
-    "webpack": "^5.72.1",
-    "webpack-bundle-analyzer": "^4.5.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.9.0"
-  },
-  "jest": {
-    "testMatch": [
-      "<rootDir>/test/unit/**/*.spec.ts",
-      "<rootDir>/design-system/test/**/*.spec.ts"
-    ],
-    "testEnvironment": "jsdom",
-    "transform": {
-      "\\.(css|less|sass|scss|styl|jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "jest-transform-stub",
-      "\\.(ts|html)$": "@aurelia/ts-jest"
-    },
-    "collectCoverage": true,
-    "collectCoverageFrom": [
-      "src/**/*.ts",
-      "design-system/**/*.ts",
-      "!src/**/*.d.ts",
-      "!design-system/**/*.d.ts"
-    ],
-    "globals": {
-      "ts-jest": {
-        "isolatedModules": true
-      }
-    }
+    "vitest-mock-extended": "^0.1.13"
   }
 }
