+++ src/resources/value-converters/index.ts
@@ -4,4 +4,3 @@ export * from './currency';
 export * from './percentage';
 export * from './ethwei';
 export * from './small-hex-string';
-export * from './timespan';
